remove the macros file
