+++ crates/js/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod bundle;
-mod macros;
 
 pub use bundle::{bundle_for_filename, bundle_hash, bundle_hash_for_filename, TsjsBundle};
